created the game over screen function and incrementing the player speed
