+++ ten/app/gamescreen.py
@@ -1,110 +1,6 @@
 from buttons import MenuButton
 import arcade
-
-PLAYER_SPEED = 0.5
-PLAYER_SCALE = 1
-
-key = {
-    119: "W",
-    97: "A",
-    115: "S",
-    100: "D",
-    105: "I",
-    106: "J",
-    107: "K",
-    108: "L",
-    65361: "←",
-    65362: "↑",
-    65363: "→",
-    65364: "↓",
-}
-
-
-class PlayerText:
-    def __init__(self, player_sprite):
-        self.player_sprite = player_sprite
-        self.text = key[self.player_sprite.get_direction(key=True)]
-        self.x_pos = self.player_sprite.center_x
-        self.y_pos = self.player_sprite.center_y
-
-    def draw(self):
-        arcade.draw_text(self.text, self.x_pos, self.y_pos, arcade.color.BLACK, 100)
-
-    def update(self):
-        self.text = key[self.player_sprite.get_direction(key=True)]
-        self.x_pos = self.player_sprite.center_x
-        self.y_pos = self.player_sprite.center_y
-
-
-class PlayerSprite(arcade.Sprite):
-    def __init__(self, idle_dict, up_frames, down_frames, left_frames, right_frames, x, y, default_direction, up_key,
-                 down_key, left_key, right_key):
-        super().__init__()
-        self.current_texture_counter = 0
-        self.idle_dict = idle_dict
-        self.up_frames = up_frames
-        self.down_frames = down_frames
-        self.left_frames = left_frames
-        self.right_frames = right_frames
-        self.center_x = x
-        self.center_y = y
-        self.scale = 0.4
-        self.up_key = up_key
-        self.down_key = down_key
-        self.left_key = left_key
-        self.right_key = right_key
-
-        self.texture = idle_dict[default_direction]
-        if default_direction == "up":
-            self.change_y = PLAYER_SPEED
-        elif default_direction == "down":
-            self.change_y = -PLAYER_SPEED
-        elif default_direction == "left":
-            self.change_x = -PLAYER_SPEED
-        elif default_direction == "right":
-            self.change_x = PLAYER_SPEED
-
-    def on_key_press_control(self, key, modifiers):
-        if key == self.up_key:
-            self.change_x = 0
-            self.change_y = PLAYER_SPEED
-        elif key == self.down_key:
-            self.change_x = 0
-            self.change_y = -PLAYER_SPEED
-        elif key == self.left_key:
-            self.change_y = 0
-            self.change_x = -PLAYER_SPEED
-        elif key == self.right_key:
-            self.change_y = 0
-            self.change_x = PLAYER_SPEED
-
-    def get_direction(self, key=False):
-        if self.change_x != 0:
-            if self.change_x > 0:
-                return self.right_key if key else "right"
-            else:
-                return self.left_key if key else "left"
-
-        if self.change_y != 0:
-            if self.change_y > 0:
-                return self.up_key if key else "up"
-            else:
-                return self.down_key if key else "down"
-
-    def update_animation(self, delta_time: float = 1 / 60):
-        if self.get_direction() == "up":
-            self.texture = self.up_frames[int(self.current_texture_counter / (10 / PLAYER_SPEED)) % len(self.up_frames)]
-        elif self.get_direction() == "down":
-            self.texture = self.down_frames[
-                int(self.current_texture_counter / (10 / PLAYER_SPEED)) % len(self.down_frames)]
-        elif self.get_direction() == "left":
-            self.texture = self.left_frames[
-                int(self.current_texture_counter / (10 / PLAYER_SPEED)) % len(self.left_frames)]
-        elif self.get_direction() == "right":
-            self.texture = self.right_frames[
-                int(self.current_texture_counter / (10 / PLAYER_SPEED)) % len(self.right_frames)]
-
-        self.current_texture_counter += 1
+from player import PlayerSprite, PlayerText
 
 
 class GameView(arcade.View):
@@ -124,6 +20,8 @@ class GameView(arcade.View):
         self.player_1_sprite_text = None
         self.player_2_sprite_text = None
         self.player_3_sprite_text = None
+
+        self.music = None
 
     def setup(self):
         self.player_sprite_list = arcade.SpriteList()
@@ -207,11 +105,19 @@ class GameView(arcade.View):
 
                 sprite.left = col_index * tile_width + x_offset
                 sprite.top = self.window.WINDOW_HEIGHT - y_offset - (row_index * tile_height)
-                print(sprite.left, sprite.top)
                 if column_value != 0:
                     self.wall_list.append(sprite)
                 else:
                     self.floor_list.append(sprite)
+
+        speed_increment = 0.5
+        time_increment = 10
+        arcade.schedule(lambda _: self.player_1_sprite.increase_speed(speed_increment), time_increment)
+        arcade.schedule(lambda _: self.player_2_sprite.increase_speed(speed_increment), time_increment)
+        arcade.schedule(lambda _: self.player_3_sprite.increase_speed(speed_increment), time_increment)
+
+        self.music = arcade.Sound("assets/MountainKing.mp3")
+        self.music.play(0.7)
 
     def setup_buttons(self):
         from menuscreen import MenuView
@@ -247,6 +153,19 @@ class GameView(arcade.View):
         self.player_2_sprite_text.update()
         self.player_3_sprite_text.update()
 
+        # Check for collision
+        for player_sprite in self.player_sprite_list:
+            if arcade.check_for_collision_with_list(player_sprite, self.wall_list):
+                self.game_over()
+
+    def game_over(self):
+        self.music.stop()
+        from gameover import GameoverView
+        game_over_view = GameoverView()
+        game_over_view.theme = self.theme
+        self.window.show_view(game_over_view)
+        game_over_view.setup()
+
     def on_draw(self):
         arcade.start_render()
         super().on_draw()
@@ -257,3 +176,7 @@ class GameView(arcade.View):
         self.player_1_sprite_text.draw()
         self.player_2_sprite_text.draw()
         self.player_3_sprite_text.draw()
+
+        self.player_1_sprite.draw_hit_box()
+        self.player_2_sprite.draw_hit_box()
+        self.player_3_sprite.draw_hit_box()
